Implemented synchronous server.js
