+++ .gitignore
@@ -1,1 +1,4 @@
 .idea
+
+# dependencies
+/node_modules
